Get OpenJPEG into Windows
This was a huge pain in the ass! The symbols weren't being visible and I
tried all kinds of stuff, eventually even maually creating a .def file.
That seemed to work, but all it really did was convince the j2k_AE
project that there was an openjpeg.dll to be found somewhere.

The solution was eventually to #define OBJ_STATIC in both
openjpeg.vcproj and j2k_AE.vcproj. I think at one point I may have been
doing that, but I needed to restart Visual Studio maybe?

Anyway, moral of the story is OpenJPEG is being very cavalier with their
__declspec stuff and it caused me some headaches.